Split Story controler up
